+++ vk_sdk/__init__.py
@@ -11,5 +11,5 @@ from .keyboard import Keyboard
 from .cmd import start_command, command, after_text_matcher, after_func
 
 __author__ = 'SPRAVEDLIVO'
-__version__ = '1.6.2'
+__version__ = '1.6.3'
 __email__ = 'admin@spravedlivo.dev'